Tweak consensus method 13 based on arma's comments
Instead of capping whenever a router has fewer than 3 measurements,
we cap whenever a router has fewer than 3 measurements *AND* there
are at least 3 authorities publishing measured bandwidths.

squash! Tweak consensus method 13 based on arma's comments

Replace the 'capped' part of the w line with an "unmeasured" keyword
with slightly different semantics.  "Unmeasured=1" means that we
didn't have enough observations for a node to use measured bandwidth
values in the authority's input, whether we capped it or not.