Partial implementation of a fix for bug 7912
I added the code to pass a destroy cell to a queueing function rather
than writing it immediately, and the code to remember that we
shouldn't reuse the circuit id until the destroy is actually sent, and
the code to release the circuit id once the destroy has been sent...

...but the code to actually implement the queueing and unqueueing of
destroy cells on the circuitmux isn't done yet.  See the big XXXXX
comment.